requirements.txt
google-auth-httplib2
google-api-python-client
google-api-core
pyyaml
